# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateSubscriptionPlanRequest:
    properties:
      description:
        type: string
      durationDays:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: integer
    required:
    - durationDays
    - name
    - price
    type: object
...
  dto.SubscriptionPlanResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      durationDays:
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: integer
      updatedAt:
        type: string
    type: object
...
  dto.UpdateSubscriptionPlanRequest:
    properties:
      description:
        type: string
      durationDays:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Api server for DaraTil PLatform
  title: DaraTil Backend
  version: "0.2"
paths:
...
  /subscription-plans:
    get:
      consumes:
      - application/json
      description: Get subscription plans with filters
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Duration days
        in: query
        name: durationDays
        type: integer
      - description: Is active
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionPlanResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List subscription plans
      tags:
      - Subscription Plans
    post:
      consumes:
      - application/json
      description: Create a new subscription plan
      parameters:
      - description: Subscription plan create request
        in: body
        name: planReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionPlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create subscription plan
      tags:
      - Subscription Plans
  /subscription-plans/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription plan by ID
      parameters:
      - description: Subscription plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete subscription plan
      tags:
      - Subscription Plans
    get:
      consumes:
      - application/json
      description: Get subscription plan by ID
      parameters:
      - description: Subscription plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionPlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get subscription plan by ID
      tags:
      - Subscription Plans
    patch:
      consumes:
      - application/json
      description: Update subscription plan by ID
      parameters:
      - description: Subscription plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription plan update request
        in: body
        name: planReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionPlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update subscription plan
      tags:
      - Subscription Plans
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"